ConversionFolder/app/.claude/worktrees/youthful-einstein/app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.22
lxml==5.3.2
anthropic==0.34.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
jinja2==3.1.6
aiofiles==24.1.0
httpx==0.27.2
itsdangerous==2.2.0
openpyxl==3.1.5
bandit==1.8.3
